Add files via upload
Project Overview
This Python project uses the PyTube library to download YouTube videos by taking the video URL as input. The project allows users to enter a YouTube URL, then downloads the video in the highest available resolution to a specified directory on their computer. It provides an easy-to-use command-line interface for downloading videos from YouTube.

Key Features
Input Video URL: The user provides the URL of the YouTube video they want to download.
Download Video: The script downloads the video in the highest resolution available using PyTube.
Save Video Locally: The video is saved to a default location (or user-defined path) on the local machine.
Libraries Used
PyTube: For fetching and downloading YouTube videos.